Rosapi moved to own crate
