+++ roslibrust_rosapi/README.md
@@ -1,0 +1,7 @@
+# roslibrust_rosapi
+
+Provides a trait representing the capabilities of the rosapi node.
+
+And provides implementations of that trait for roslibrust_ros1::NodeHandle and roslibrust_rosbridge::ClientHandle
+
+The future of this crate is a little uncertain given direction of the roslibrust project.
